improved input system, cleaned up some notes, & removed unused stuff
inputs can now be a number between 0 and 100. It used to just be 10-99, also it can technically go up to 999, but the program was not designed for that so idk what would happen as I have not tested that, will probably add an input validation at some point. 
The new input system is still not perfect though. If you are to input a 1 digit number, the hundreds and tens place MUST have a zero in them, and the same goes for 2 digit numbers (example: to enter 5, you must type 005).
Also removed the TEN & NBASE labels, since they are no longer used